requirements.txt
streamlit
pandas
networkx
requests
python-dotenv
openai